requirements.txt
aide-ds @ git+https://github.com/Asimawad/aide-ds.git 
openai==1.65.0
python-dotenv==1.0.1
psutil==5.9.8
